requirements.txt
Django==5.2.4
asgiref==3.9.1
sqlparse==0.5.3
psycopg2-binary==2.9.10
djangorestframework==3.16.0
drf-yasg==1.21.10 